PR Feedback - split up views and data controllers into separate files
