+++ WMFComponents/Sources/WMFComponents/Components/Year in Review/View Models/WMFYearInReviewLocationViewModel.swift
@@ -1,0 +1,10 @@
+struct WMFYearInReviewSlideLocationViewModel {
+    
+    let title: String
+    let loggingID: String
+    
+    init(localizedStrings: WMFYearInReviewViewModel.LocalizedStrings, loggingID: String) {
+        self.title = localizedStrings.locationTitle
+        self.loggingID = loggingID
+    }
+}
