+++ WMFComponents/Sources/WMFComponents/Components/Year in Review/Views/Shareable Slide Views/WMFYearInReviewSlideStandardShareableView.swift
@@ -1,21 +1,24 @@
 import SwiftUI
 
-struct WMFYearInReviewShareableSlideView: View {
+struct WMFYearInReviewSlideStandardShareableView: View {
+    let viewModel: WMFYearInReviewSlideStandardViewModel
 
     @ObservedObject var appEnvironment = WMFAppEnvironment.current
 
-    var theme: WMFTheme {
+    private var theme: WMFTheme {
         return appEnvironment.theme
     }
 
-    let imageName: String
-    var altText: String
-    var slideTitle: String
-    var slideSubtitle: String
-    var hashtag: String
+    private let hashtag: String
+    
+    init(viewModel: WMFYearInReviewSlideStandardViewModel, appEnvironment: WMFAppEnvironment = WMFAppEnvironment.current, hashtag: String) {
+        self.viewModel = viewModel
+        self.appEnvironment = appEnvironment
+        self.hashtag = hashtag
+    }
     
     private var attributedString: AttributedString {
-        return (try? HtmlUtils.attributedStringFromHtml(slideSubtitle, styles: styles)) ?? AttributedString(slideSubtitle)
+        return (try? HtmlUtils.attributedStringFromHtml(viewModel.subtitle, styles: styles)) ?? AttributedString(viewModel.subtitle)
     }
     
     private var styles: HtmlUtils.Styles {
@@ -32,7 +35,7 @@ struct WMFYearInReviewShareableSlideView: View {
                         .foregroundColor(Color(theme.text))
 
                     ZStack {
-                        Image(imageName, bundle: .module)
+                        Image(viewModel.gifName, bundle: .module)
                             .resizable()
                             .scaledToFit()
                             .frame(maxWidth: .infinity)
@@ -41,7 +44,7 @@ struct WMFYearInReviewShareableSlideView: View {
                     }
                     .padding(.top, 10)
                     VStack(alignment: .leading, spacing: 12) {
-                        Text(slideTitle)
+                        Text(viewModel.title)
                             .font(Font(WMFFont.for(.boldTitle1, compatibleWith: UITraitCollection(preferredContentSizeCategory: .medium))))
                             .foregroundStyle(Color(uiColor: theme.text))
                         Text(attributedString)
