+++ WMFComponents/Sources/WMFComponents/Components/Year in Review/Views/WMFYearInReviewDonateButton.swift
@@ -13,7 +13,6 @@ struct WMFYearInReviewDonateButton: View {
     var body: some View {
         Button(action: {
             viewModel.handleDonate(sourceRect: buttonRect)
-            viewModel.logYearInReviewDidTapDonate()
         }) {
 
             ZStack {
