+++ WMFData/Sources/WMFData/Data Controllers/Year In Review/Slide Data Controllers/YearInReviewEditCountSlideDataController.swift
@@ -1,0 +1,161 @@
+import CoreData
+
+final class YearInReviewEditCountSlideDataController: YearInReviewSlideDataControllerProtocol {
+
+    let id = WMFYearInReviewPersonalizedSlideID.editCount.rawValue
+    let year: Int
+    var isEvaluated: Bool = false
+    static var containsPersonalizedNetworkData = true
+    
+    private var editCount: Int?
+
+    private let username: String?
+    private let project: WMFProject?
+    
+    private let yirConfig: YearInReviewFeatureConfig
+    private let service = WMFDataEnvironment.current.mediaWikiService
+    
+    init(year: Int, yirConfig: YearInReviewFeatureConfig, dependencies: YearInReviewSlideDataControllerDependencies) {
+        self.year = year
+        self.yirConfig = yirConfig
+        self.username = dependencies.username
+        self.project = dependencies.project
+    }
+
+    func populateSlideData(in context: NSManagedObjectContext) async throws {
+        guard let username, let project else { return }
+        
+        guard let startDate = yirConfig.dataPopulationStartDateString,
+              let endDate = yirConfig.dataPopulationEndDateString else {
+            return
+        }
+        
+        let (edits, _) = try await fetchUserContributionsCount(username: username, project: project, startDate: startDate, endDate: endDate)
+        
+        editCount = edits
+        isEvaluated = true
+    }
+
+    func makeCDSlide(in context: NSManagedObjectContext) throws -> CDYearInReviewSlide {
+        let slide = CDYearInReviewSlide(context: context)
+        slide.id = id
+        slide.year = Int32(year)
+
+        if let editCount {
+            slide.data = try JSONEncoder().encode(editCount)
+        }
+
+        return slide
+    }
+
+    static func shouldPopulate(from config: YearInReviewFeatureConfig, userInfo: YearInReviewUserInfo) -> Bool {
+        return config.isEnabled && config.slideConfig.editCountIsEnabled && userInfo.username != nil
+    }
+    
+    func fetchUserContributionsCount(username: String, project: WMFProject?, startDate: String, endDate: String) async throws -> (Int, Bool) {
+        return try await withCheckedThrowingContinuation { continuation in
+            fetchUserContributionsCount(username: username, project: project, startDate: startDate, endDate: endDate) { result in
+                switch result {
+                case .success(let successResult):
+                    continuation.resume(returning: successResult)
+                case .failure(let error):
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    func fetchUserContributionsCount(username: String, project: WMFProject?, startDate: String, endDate: String, completion: @escaping (Result<(Int, Bool), Error>) -> Void) {
+        guard let service = service else {
+            completion(.failure(WMFDataControllerError.mediaWikiServiceUnavailable))
+            return
+        }
+
+        guard let project = project else {
+            completion(.failure(WMFDataControllerError.mediaWikiServiceUnavailable))
+            return
+        }
+
+        // We have to switch the dates here before sending into the API.
+        // It is expected that this method's startDate parameter is chronologically earlier than endDate. This is how the remote feature config is set up.
+        // The User Contributions API expects ucend to be chronologically earlier than ucstart, because it pages backwards so that the most recent edits appear on the first page.
+        let ucStartDate = endDate
+        let ucEndDate = startDate
+
+        let parameters: [String: Any] = [
+            "action": "query",
+            "format": "json",
+            "list": "usercontribs",
+            "formatversion": "2",
+            "uclimit": "500",
+            "ucstart": ucStartDate,
+            "ucend": ucEndDate,
+            "ucuser": username,
+            "ucnamespace": "0",
+            "ucprop": "ids|title|timestamp|tags|flags"
+        ]
+
+        guard let url = URL.mediaWikiAPIURL(project: project) else {
+            completion(.failure(WMFDataControllerError.failureCreatingRequestURL))
+            return
+        }
+
+        let request = WMFMediaWikiServiceRequest(url: url, method: .GET, backend: .mediaWiki, parameters: parameters)
+
+        service.performDecodableGET(request: request) { (result: Result<UserContributionsAPIResponse, Error>) in
+            switch result {
+            case .success(let response):
+                guard let query = response.query else {
+                    completion(.failure(WMFDataControllerError.unexpectedResponse))
+                    return
+                }
+
+                let editCount = query.usercontribs.count
+
+                let hasMoreEdits = response.continue?.uccontinue != nil
+
+                completion(.success((editCount, hasMoreEdits)))
+
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
+    
+    struct UserContributionsAPIResponse: Codable {
+        let batchcomplete: Bool?
+        let `continue`: ContinueData?
+        let query: UserContributionsQuery?
+
+        struct ContinueData: Codable {
+            let uccontinue: String?
+        }
+
+        struct UserContributionsQuery: Codable {
+            let usercontribs: [UserContribution]
+        }
+    }
+
+    struct UserContribution: Codable {
+        let userid: Int
+        let user: String
+        let pageid: Int
+        let revid: Int
+        let parentid: Int
+        let ns: Int
+        let title: String
+        let timestamp: String
+        let isNew: Bool
+        let isMinor: Bool
+        let isTop: Bool
+        let tags: [String]
+
+        enum CodingKeys: String, CodingKey {
+            case userid, user, pageid, revid, parentid, ns, title, timestamp, tags
+            case isNew = "new"
+            case isMinor = "minor"
+            case isTop = "top"
+        }
+    }
+
+}
