+++ WMFData/Sources/WMFData/Data Controllers/Year In Review/WMFYearInReviewUserInfo.swift
@@ -1,7 +1,7 @@
 import Foundation
 import CoreData
 
-public struct YearInReviewUserInfo {
+struct YearInReviewUserInfo {
     let username: String?
     let userID: String?
     let project: WMFProject?
